Fix call to PUT artifact_temporary files
request #11720 Error: Bad Request: `content` is required in the artifact
modal
Also needed for request #11719 File upload with the artifact modal
freezes the browser

Since a refactoring of the rest-service, the route PUT
artifact_temporary_files/:id is not called with the correct headers and
body anymore, therefore returning Bad Request.

How to test:
- Try to attach a file size > 1 MiB. It should succeed. Without this
patch, there would be an error.

Change-Id: I8abc03d9d3bd2929368c85140d1d8700b96f3ad6