Remove completly the mention of the gitphp-tuleap and Smarty packages
These packages have been marked has obsoleted and conflicted with
plugin-git and plugin-git-gitolite3 in order to remove them since they
are not needed anymore.
This confuses RPM during an upgrade.

Also, the dependency to GeSHI is provided by Composer so it is
not needed to install it from EPEL anymore.

This part of story #10411: cleanly integrate gitphp into tuleap

Change-Id: I6d3d9debe17151b7608e5e03c4db4c9cd38a7727