Controllers can inform about the current project
There is an automated control of private project / restricted users made by
URLVerification in tuleap based on 'group_id' parameter in URL.

Until now, when no group_id was provided, an event was sent to let implementors
provide the id if they recognize an URL intended for them.

It both complex (group_id automagically generated somewhere) and inefficient as
the URI is parsed several times to finally be associated to the right plugin.

With this contribution, a controller associated to a Route can return the
project it's working on and bypass the event stuff.

Part of story #10408 have a burning parrot git repository list

Change-Id: I1693974d6c6299b93c738fcffb89bd1ca1adafe4