Use the XML in archive mode in move artifact
This is part of story #11346 move artifact's "initial_effort"

No functionnal changes here. Given that we will move a field that can be
in autocompute mode, we have to move the computed value in the last
changeset and not the autocompute mode.

This change will allow us to do that in the next commits.

Change-Id: I9232bfa72a7d3961f8d327d7988fb80da24acbc1