SemanticChecker checks for external semantic
This is part of story #11346 move artifact's "initial_effort"

Now the initial effort semantic is checked before the move operation.
If no tracker semantic are aligned (title and description for now) we
will ask to other plugins if their semantics are aligned (in our case
the Agiledashboard plugin and the initial effort semantic).

The initial effort semantic is considered aligned if both tracker have
the semantic and if the field have the same type.

The initial effort semantic value is not move yet, it will be done in
the next commit.

Change-Id: I84a36e0f3ad4b49ce36bca18a52ce0150b747ee0