Fix margin for git cards
Part of story #10408 have a burning parrot git repository list

Go to /plugin/git/<projectname>/

The margin on the right of the right column should be the same than on the
left.
If you have an odd number of repositories, then the last one should not span on
two columns.
If you resize your browser, cards should be on one column only.

Change-Id: Iad87fa1fb6ee61e6ff82ce126592d5bfdced7fcc