Display selectbox to see user repositories
Part of story #10408 have a burning parrot git repository list

Given a project with forked repositories,
When you go to /plugins/git/<project/,
Then you see a selectbox that contains the users who forked a repository.
And for project where there isn't any forked repository,
Then the selectbox is not displayed.

This selectbox does nothing for now. Display of selected user's repositories
will be done later.

Bonus: fix also the style of the empty state: the primary button should be
large and shouldn't be duplicated

Change-Id: Iff79d50168f486f9340e91d1a0e4a59303bc1eef