Stop (re)writing $_GET and $_REQUEST variables to display the GitPHP view
The remaining rewriting for "pre-sanitization" purposes is kept
for now. It will be dropped once all the unused templates are removed
to reduce the needed work to check if those variables are always properly
escaped.

The tree view is now the default view instead of the summary view.

This is part of story #10411: cleanly integrate gitphp into tuleap.
It is also part of story #10415: see tree view by default.

Change-Id: If4eb7680311b50d166e437f809b47be0e18c61e0