Selecting a user displays their forked repositories
Part of story #10408 have a burning parrot git repository list

Given a project with forked repositories,
When you go to /plugins/git/<project/,
Then you see a selectbox that contains the users who forked a repository
And if you select a user
Then the list of repositories is changed to display their forked repositories,
And if you select back the project repositories option
Then the list of repositories is changed to display the project repositories.

The REST calls are cached to help saving the rain forest.

Change-Id: I0b1ce4ce30229c965a2ac155feb8265a990fedff