Enable PDO by default on all instances
In case of issue it is still possible to disable it with the
setting fallback_to_deprecated_mysql_api.

Also note that a few parts of Tuleap are still using the mysql_* API
such as PHPWiki or the setup script of the IM plugin.

This is part of story #11146: Provide PDO by default for all Tuleap DB query

Change-Id: I134c7739047db52a64663032cb47bd56b22a5723