request #11784: Pull request can not be accessed anymore if the Git history is rewritten
Tuleap must have the right to rewrite the history in the namespace refs/tlpr
otherwise the pull request reference can not be updated if the history
of the pull request is rewritten too.

Say otherwise, we should still be able to access a PR if we have force push
in the PR branch.

Change-Id: I1e6db7dba745f5d8581dfeff1b34bb8a8b5d2c67