request #11796 Too many SQL calls on "plugin" table done at each hit
When profiling a simple page like the new git repositories listing, ~50 of the
queries (19 on 40) are related to the plugin table (to know the name of one
specific plugin or to know if a plugin is restricted or not).

It's more efficient to make a single call for all plugins and to inject the
informations when it's instanciated.

As plugin table is small and probably utterly cached by mysql, the gain is
almost null in the whole process but that's still reduce mysql pressure.

Change-Id: I5b9292e7809680834f0844bfba9632f204fbd563