request #11565 : Modify the performance REST route /git/{id}/statuses/{commit_reference}
This REST route has been modified because it may be hard for an user to retrieve the git id repository on Tuleap.
The end user does not have to search the ID repository which is hard to have.
It will also make the Jenkins build configuration easier.
The new version accepts also the git repository path and project name.

How to test :
- Create a git repository with at least one pull request.
- Go to the Tuleap API Explorer page.
- Go to the « git » section
- Go to /git/statuses/{commit_reference}
- Fil all the field. (*)
- Click on « Try it out!»
- If you filled correctly all the field, it should return 201.
- Go to the pull request and you should see if your last build successed or failed.

(*): id_or_git_path format => <Project name>/<git repository path>.git

Change-Id: Ice37504d3cbc39f1b9a311433abd579d87d2d6d1