Dedicated query for counting all git repositories
Creation of GitRepository objects is very expensive because they load all email
notifications, it's 1 SQL query per repo but for 1000 repo => 1000 queries.
It's already bad on the  principle but those 1000 queries are executed even if
only 10 repo are listed thanks to limits.

On small projects the gain is modest (5% of wall time for 13 repositories) but
 it's still that.

Part of story #10408 have a burning parrot git repository list

Change-Id: I33fd493d3d48f0e43f1a72daa8217f2c2bad912c