request #12475 [Perf] cache hooks and plugins in a more aggressive manner
Let's start with some data, compare to master on empty.php
- Wall time: 34.9 ms -> 21.6 ms (-38%)
- Memory: 9.23 MB -> 7.28 MB (-21%)
- SQL queries: 8 -> 7

It saves a lot of time because there are far less class to instantiate, enhance
autoloads to kick.

On other, Burning Parrot pages, the pre.php part seem down to ~8.5ms and we
can see a 13% perf gain.

You need to activate it by setting in local.inc
$plugin_hooks_cache_type = 'serialized';

Change-Id: I168361078590cf913ea139bb3200a515e781d3be