Repos list should be retrievable even if a a repo has been disconnected from Gerrit
A Git repo that has been disconnected from a Gerrit server keep
the remote server ID even if the Gerrit server has been deleted.
Also PHP evaluates -1 as true when checked in a condition so it breaks
when we try to build the server representation when we try to build the
server representation from the cache data.

This is part of story #10408: have a burning parrot git repository list

Change-Id: Iad4d76a6f0185cd2084d94369160b75749fe8243