Errors might be hidden when if the user is redirected to the Git plugin homepage
When a user is redirect from the "old" Git home URL to the new one,
the feedback messages are not re-serialized to be displayed after the
redirection so they are lost.
It is usually not an issue because only users try to access the redirected URL
so there is no feedback messages to save, here it is a bit different because
some parts of Tuleap still redirect to it.

The issue can be reproduced by trying to migrate a repository to Gerrit
without choosing a permission's configuration template.

This is part of story #10408: have a burning parrot git repository list

Change-Id: I653d638d108758d0b7f34898615f7ac34935d38c