request #11848: Kanban widget can not be deleted
On most instances the deletion worked before 0d72d085de4eb0889c926e7bc6ea584f08184b2e
because a transaction is started on the mysql driver connection and another
transaction was started on the mysqlnd driver connection. Now by default
all interactions with the DB are done with mysqlnd driver so we try
to start a transaction within a transaction. Nested transactions are (thankfully)
not supported by MySQL so we get an error.

Change-Id: I8058ddb83cbf97bbd5cf6d4b5e1e77512e11ede1