Enrollment of a multi-factor authentication device based on TOTP can be saved
Even if it is not used yet, this contribution saves the enrollment of a
MFA device in the DB.

To test, you can access the page /plugins/mfa/enroll a base32 encoded value
is displayed, it is the shared secret between the client and the server
validating the TOTP codes. You can generate the TOTP code corresponding to the
secret using the CLI tool oathtool (any other tool will work too) this way:

    oathtool --totp -b <secret>

The plugin not being not available outside of a dev environnement and the
contributor being lazy, no forgeupgrade is provided so please create the table
by hand or just uninstall then reinstall the plugin.

This is part of story #11649: protect my account with a TOTP code

Change-Id: I13000cc56ca6eca75bd83e7164700768dcbe607b