Refactoring: Each button should have its own mustache file
this is part of story #11347 have a UI to accesss Move artifact

This commit split the way presenters are built, the aim is to migrate
the "move" button a the entry point of vue app.
In order to ease review, I'll do the button as view app in a dedicated commit.

No functional changes, all button shoul be displayed as usual

Change-Id: If9ff06f6b829973df8380590f83b4cafc270aa98