Setting to only merge PR in fast-forward can be set in database
The setting is not yet taken into in the UI or during fork and project
creation. However if you disallow merge commits for a repository and try
to merge a PR needing a merge commit, the merge will be rejected with a 400.

This is part of story #11661: define a merge strategy for my repositories

Change-Id: Ie30eded79be088cf2f36d8e5fde3cf969482317c