Merge commit allowed option is added for template
This is part of story #11661: define a merge strategy for my repositories

This option is not yet visible through web UI.
This option is taken into account when creating a new repository (not
the forked ones).

The inheritance of this template option and the web UI will be done in dedicated
commits.

Change-Id: I65db8b7a93f17f098c3b76442053eff86d236ba8