Create a dry-run mode form PATCH artifact/:id
This is part of story #11347 have a UI to accesss Move artifact

A boolean attribute "dry-run: true" can be provided in the PATCH body content
to have a feedback of the move without doing the action.

This feedback contains fields that will be moved and fields that not
will be moved based on the semantics.

Change-Id: I43e6f40e286219bbc1b7625703814208074ccd23