Refactoring: move display of default settings
Part of story #11661: define a merge strategy for my repositories

In order to be able to add a new pane in this view, and to not lose my mind, I
propose to extract the display in a dedicated object, so that we don't have to
play to much with Git(|Actions|Views) god objects.

No functional changes, except that if you perform an action in default
settings, then you will be redirected after the action (which should always be
the case after a POST).

Please note that only the display of default settings is impacted, handling of
actions (access controls & mirrors) is still done by Git/GitActions. That will
be done later. Maybe.

Change-Id: Ifdb6c1d574c905fbc06ab9d2ae62f56cd2b615fa