Refactoring: decrease responsibility of IndexController
Part of story #11661: define a merge strategy for my repositories

IndexController should not have the responsibility to retrieve mirrors or
permissions. Its current duty is to only display relevant pane. Retrieving
mirrors or permissions is now delegated to dedicated Pane objects. This will
help to introduce Pull requests pane.

Took the time to add missing disabled panes.

An issue introduced by this commit is the introduction of {{{ }}}. However,
since there is no way to specify dynamically (by the presenter) which partial
to load (because PR is a plugin and we cannot leak PR content in Git
templates), then I do not see a better approach.

Change-Id: I117a511ccf0ff5d774323070befff6e35722c7ba