request #12067: Move to the remi-safe repository for the PHP dependencies
When upgrading users will be expected to redeploy the PHP-FPM configuration
with '/usr/share/tuleap/tools/utils/php56/run.php --module=nginx,fpm' and
to stop the existing rh-php56-php-fpm service.

If an existing PHP-FPM configuration is found for rh-php56-php-fpm, the exact
same configuration will be redeployed for php56-php-fpm in order to keep
customizations that may have been done by administrators.

This contributions has led to the rebuild of a lot of the Docker images
and subsequently to some clean up of the tests.

The antialising of JPGraph has been disabled until we can bump JPGraph
version. The php-gd extension provided by the remi-safe package does not
rely on the version of the library provided by PHP but on the one provided
by the system which does not have the antialiasing feature.

Change-Id: I286d375ace6f0596154710e603a3db4431329cbd