request #12082: Usage of old themes can lead to issues
The theme value that may have been set per users in the past
is dropped. Tuleap only supports theme variants, this value
is not needed anymore.

Also since the introduction of BurningParrot, running a custom
theme is not really possible anymore. All the logic around that
is dropped as well.

Change-Id: I0b83652d69b70389eb362e4437f05a171f097b8e