Add the REST route GET /git/:id/files?path=path/to/file&ref=string
This is a part of the story #11419.

The goal of this new REST route is to improve the performance of the future Jenkins plugin Tuleap Branch source.

How to test :
- In your git respository, put some files.
- Go to the api explorer
-  Fill the fields.
- Click on ‘Try it out !’

Expected results :
If all is ok => 200
If the ref is not good => return 404
If the path of the file is not good => return 404

Change-Id: Idcba9a9ffcf774fdf1ccd15ccb589e642d827938