request #12079: Drag&drop of cards in cardwalls or columns in tracker reports generate JavaScript warnings
We explicitely disable the passive behaviors of the touch events [0]
in order to keep the behavior expected by the drag&drop library and
stop browsers from complaining.

This is of course not a proper fix but doing something as it should
be done nowadays would probably take too much time with a library
still containing fix for IE6.

[0] https://developers.google.com/web/updates/2017/01/scrolling-intervention

Change-Id: I26ad65f9f2c0d5ac22868de236a83f238e8269fe