Do not display preview title when there is no preview
this is part of story #11347 have a UI to accesss Move artifact

There should not be a preview state in case of error. Try to move an artifact
in the same tracker: header doesn't change. Furthermore, project and tracker
selectors are not reset so that the user can fix their error.

Change-Id: I75fb2998c9822fcc349834df349725a67069635e