Import the Move artifact modal dynamically
Part of story #11347 have a UI to accesss Move artifact

How to test:
- build the app
- Open the "Move artifact" modal. It's javascript code should be
lazy-loaded and the "Actions" dropdown icon should show a spinner.
- No functional change

Note: the error feedback is not translated, but I'm not sure it's worth
it to include node-gettext just for cases where a file cannot be loaded
from the server.

Change-Id: I92b92337fa0c27bb9730f72fab0c6cc265c834b1