story #11842 filter cardwall columns
Given a cardwall tracker report with a list field selected.
If the report filters on this field in classic search,
then the cardwall columns will only be the selected values.

This must works with all type of list values (static values, users and
ugroups).

This change is also reflected in dashboard. Nothing changes
for the AD cardwall.

The TQL search is not taken into account. While switching from classic
to advanced search (or vice versa) you have to save your report to have
(or not) filtered columns.

Change-Id: Ie0666cd72d6dd1c87daf015b734c4fdde5475c7c