Add a feature flag and have Burning parrot repository name
This is a first step to allow progress on the switch to BurningParrot
for Pull requests and Git.

How to test:
- Browse to a repository. No functional change expected in FlamingParrot
- add $git_repository_bp = 1; in you development.inc to switch to
BurningParrot
- Using BurningParrot, only the repository's name and path and a Setting
button can be seen in the header page.

Will be done later:
- GitPHP styling (margins and font-sizes changed)
- The rest of the header
- Removing FlamingParrot-related code

Note: Since normal BP pages must use DispatachableWithBurningParrot
(implements), we use a hackish way to use a flag to switch to FP or BP.
It lets us split the dev into small commits.

Change-Id: I44f05cfe1d8fb829cd7324f021ed9ed77a067deb