tasks #12094 Create modal to get informations in plugin Tuleap_synchro second part of three of request #12084 Add form in plugin Tuleap_Synchro
Create a form in the modal in the plugin tuleap_synchro
All informations in this form, will (after) store in database, and use to launch an automatically update between 2 tuleaps' instance.
Currently, when we click on "add endpoint", nothing to create, and we return to the plugin.

Project's shortname must be just shortname (not id).

How to test :
	Put information in the form

Expected results :
    	If base URI doesn't comply, a message says this.
	If the tracker isn't a numeric, a message says this.

	We can't valid, the button is disabled

Change-Id: I3dd969a6ed8c83d8c79a0438a8f84c0a33d2c2ad