Fix encoding issue with french feedback
Part of story #11347 have a UI to accesss Move artifact

When you successfully move an artifact, feedback should not have encoding issue
like "a ?t?" instead of "a été".

The bug is due to usage of setlocale to getlocale (sic). According to the
documentation[1], one should passe "0" and not null to retrieve the current
locale.

Also add the missing %s in translation.

[1] http://php.net/setlocale

Change-Id: I33b1cd0678c27851377792a973180354b8a75a3e