Refactoring: Fix mutations and actions
Part of story #11347 have a UI to access Move artifact

No functional changes expected.

- Add jasmine 3.2's async matchers for rest querier tests. It reminds of
the jasmine-promise-matchers plugin we added for AngularJS and helps a
little with asserting native promises.
- Reworked the mutations. I'm sick of getter/setter in Vuex mutations.
No more! Now we have meaningful operations. States that are changed
together should be in the same mutation. Booleans are "reset" when
switched to false. As a result, there is a lot fewer mutations.
- Actions do the redirection. Since "window" can't be spied upon by
jasmine, we need a little helper module that we can then rewire.
Business as usual. As a result, there is no more "should_redirect" flag.

Change-Id: Ib6607684b6e6e456d7377c1d3eb1174b083d3377