request #12118: Mails with only BCC are rejected
The issue can be seen when starting a new thread in the forum
service or expecting a mail from the COMPUTE_MD5SUM system event.
The issue has been introduced/made visible by commit
a9a25adfa40b575f9e6487d603a6723ab89674bb.

If everything was working as expected we should not even have to
check that, Zend\Mail should handle it properly and report an issue
if no To, CC or BCC header is found. Unfortunately, due to a bug [0]
in the version we currently use if no To header is set Zend\Mail complains
even if we have a BCC header. This contribution work around that until
we are able to bump our dependencies to Zend.

[0] https://github.com/zendframework/zend-mail/issues/73

Change-Id: I14b3f4f3d21ec89ef59a4fd127f4ab0c98ef68da