request #12132: Run (almost) all the SimpleTest tests in PHP 7.2
A few tests are not yet run in in PHP 7.2:
* ForumML_InsertTest: does not run because of the dependency on
PEAR Mail_Mime. This will be solved by the ongoing work in
request #12126.
* multiple tests in the OpenID Connect plugin: a few classes does
not respect the method signatures when redefining methods. This will
be solved with the fix for request #9527.
* integration tests of the project import: our version of Mediawiki
and the associated plugins are not compatible with PHP 7.2. Due to the
scope of the required changes to make it work, it will be fixed in a
dedicated request/story. However, besides Mediawiki, the project import
feature is working in PHP 7.2 as it is already used in the REST and SOAP
tests.

Change-Id: I5288992050192818076756d0fd2254d5f8e69fae