Tracker email gateway is compatible with PHP 7
The usage of PEAR Mail_Mime is replaced with php-mime-mail-parser.
While the latest version of PEAR Mail_Mime is announced compatible
with PHP 7.2 it can only be found on PEAR.
To ease future maintenance a library available on Packagist is now used.
It avoids the need of repackaging PEAR Mail_Mime or to maintain a
version of PEAR Mail_Mime that can be retrieved with Composer.
php-mime-mail-parser depends on the PHP extension mailparse which
is generally available in the repositories of distributions
(at least it's the case in RedHat/Fedora and Debian/Ubuntu worlds).

This is part of request #12126: Move away from PEAR Mail_Mime library

Change-Id: I15c8fc46ef76dd50cab947e38bce1184863d58af