request #12150: Reduce load pressure put on CI agents running Tuleap test suites
This contribution is a major reorganization/rewrite of the Tuleap test pipeline:
* running all tests in parallel has become inefficient, some test suites has
been grouped together to be executed sequentially. The execution of the
grouped test suites is expected to be quick enough to run while longer
test suites are played.
* REST tests do not copy the Tuleap sources before lauching the test suite.
REST tests do not write anything anymore in /usr/share/tuleap, it is safe
to use the same checkouted and prepared source folder than the other
tests.
* actions needed to be done to launch test suites or checks has been
extracted from the pipeline. It is a step toward pipeline reusability
for, in particular, external plugins: a minor change in a way a test
suite is launched will not result in a complete failure of the outdated
pipeline like it was the case until now. It also reduces the amount of
copy/paste that is needed create a new pipeline.
* code coverage for the PHPUnit test suites has been disabled. It is
slow and there is no evidence that Tuleap contributors or maintainers
use it this way and nothing is enforced as of today. Contributors can
still launch it in their local environnements to help them during their
developments or code reviews. However, code coverage is an interesting
metric that could be computed in our newly nightly and monthly
pipelines.

Change-Id: I0f1a3891924a54f70b4056149a9321877721d664