request #12149: Bump Kanban dependency to socket.io-client ^2
The realtime is expected to continue to work as before.

While being a major version upgrade to a breaking change [0][1],
not a lot of trouble is expected from this contribution.
Even if our server still uses socket.io ^1, the TestManagement plugin
is using socket.io-client ^2 since a while without particular issues.

[0] https://github.com/socketio/socket.io/releases/tag/2.0.0
[1] https://github.com/socketio/socket.io/issues/3007#issuecomment-336791836

Change-Id: Ic7f3fd4f9acacfc3fd5ddd48c82523f73f6495c6