Enable button [Create PR] if there are branches
Part of story #10409: have a burning parrot git repository view

Go to a repository with branches: button should be enabled.
Go to a repository without branches: button should NOT be enabled.
Go to a fork with branches: button should be enabled.
Go to a fork without branches but with branches in the parent repository: button should be enabled.
Go to a fork without branches and without branches in the parent repository: button should NOT be enabled.

Change-Id: Ice5533191c12c6c2cc59c0efaa1f52640764c2ca