request #9527: OpenID Connect plugin does not work with Keycloak
The issue is caused by the fact that Keycloak adds a parameter
not specified by the specification [0][1]. This parameter name
contains a dash which make the library currently used by Tuleap
to process the token crash when the fields of the token response
object are generated.
This contribution replaces the library usage to process the token
request and response.

[0] https://openid.net/specs/openid-connect-core-1_0.html#TokenResponse
[1] https://tools.ietf.org/html/rfc6749#section-1.4.1

Change-Id: I62724dab13f6cc784144d866209b02f7e0ffbb1b