Refactoring: split files
Part of story #10409: have a burning parrot git repository view

In order to be able to introduce the modal without having only a big root component, split now the root component.

Rename the root component as App.vue so that we instantly know what is the entry point when we browse files.

No functional changes.

Change-Id: Iabd64394af93632e6171618e8a9726a5877cafee