Display modal to create the pull request
Part of story #10409: have a burning parrot git repository view

When the [Create pull request] button is enabled, if user clicks on it a modal
is displayed so that branches (source and destination) can be chosen to create a
pull request.

The pull request is not created yet, it will be done in a next commit.

Change-Id: Ibffb7e873191c4650a4e41efd6f81ae84e791040