Refactoring: Diff directive clean-up
Part of story #11667 display diff side by side

- When you try to submit an empty inline comment, your browser should
stop you and tell you the textarea is required.
- When you submit an inline comment, a spinner icon is shown on the
"Comment" button.
- No other functional changes expected.

I took the opportunity to rearrange file-diff's code in the controller
instead of the directive's link function.

Also upgrade CodeMirror version. It was pinned because we needed
CodeMirror to be present on window. Now that we use webpack and import
it, we no longer care \o/

Change-Id: I3bc09520d339a1efb91ef8cf3cc3243209f5aa82