Bypass repository selection when I fork
Part of story #10409: have a burning parrot git repository view

In breadcrumb, choose repositories » fork. You have a first page that let you
select a repository, the namespace, the scope and destination project. Then you
have a page to set the permissions.
Now go to a repository, click on Fork button: the first page is bypassed and
you can directly set the permissions. Repository is created as a personal fork
in the namespace u/<username>/<repository>.

Change-Id: Ic6e3d4cfdf5c2fb7761fe1dd86089949788a1095