Display readme in a tlp-pane
Part of story #10409: have a burning parrot git repository view

Go to a repository with a README.mkd file. The content of the file should be
displayed in its own pane and the title should not look like broken.

We do not display the pane on the right like in the mockup yet. It will be done later.

Change-Id: Ic7e3d163632bdffcbaaace87fe365182d6f68444