Close modal after Pull Request creation
Part of story #10409: have a burning parrot git repository view

Go to pull request dashboard for a repository. Create a pull request. You
should be redirected to the new pull request and the modal should be closed.

Note: while the browser is reloading the page, the angular app catch the #hash
change in the url and perform its routing. We can see it behind the backdrop. I
didn't test it, but maybe in another commit we could send a custom event when
we reload the page, so that the angular app does not route?

Change-Id: I83255f1a1022ddcb0300c7f21376b5336c5568a6