request #12223: Openlist bound on users should not load all the users when no default value is defined
If you have an OpenList field with a bind on large group of users (e.g. registered users)
without a default value, the whole list of users is retrieved when you get the structure
of the field.

With a large number of users in the group the operation takes too much, not needed,
CPU time and memory.

This is particularly visible on tasks needing to retrieve the tracker such as the
artifact modal in the agile dashboard.

Change-Id: I12fee7abc1f701232f69d47414df4cbc36436aef