Fix fatal error on pull request routing
Part of story #10409: have a burning parrot git repository view

Go to the PR dashboard.
Load closed PRs.
Go to a PR overview.
Click on the tab "Pull requests (x)"

You should not have an error modal.

Angular should not make subsequent calls when we click on the tab, but this
will be fixed in a dedicated commit.

Change-Id: I09fe18b282f5b7cc47de757757ceae99d56f7d0a