User information are retrieved without the help of an external lib
Tuleap now manages directly the retrieval of data from the UserInfo endpoint.
This contribution reduces the dependency to an external library, the goal is
to remove it completly to be able to drop completly the non compatible with
PHP 7.2 code that is currently used.

request #12235: Drop usage of InoOicClient library in the OpenIDConnect client plugin

Change-Id: Ifacb2850385afd44dc62563fe0f71a74df082e7c