List commits of a pull request
Part of story #11664: display list of commits of a pull request

Go to git » repository » pull requests » pr » commits.
You should see the list of commits of the pull request.
If you click on a commit, you are redirected to gitphp view.

Commits are not grouped by date yet.
Build status badge is not displayed yet.

Change-Id: I454a8ca83617356de29c9342bfa0ba4cb5d18d61