Reduce the number of SQL queries needed to display an artifact links table
This contribution only tackle what is done in the field artifact link itself.
There is still a lot of optimizations that can be done in rendering of the tracker
report itself but it will be handled in a dedicated contribution.

This is part of request #12245: Displaying an artifact with a lot of artifact links is slow

Change-Id: I2e880d8d0cff04d0674fea6c378ca8a5cdf4c8c7