Improve performances of the tracker report renderer table when displaying artifact links
This contribution works on two fronts:
 * reducing the number of SQL queries
 * caching results that may be used multiple times to avoid recomputing
it each time

While it would be preferable to separate the part doing the caching and the
part really dealing with data but the current architecture of the tracker
form elements does not allow to do this kind of changes safely. Therefore
while that is not ideal, this contribution uses local caches quite conservatively
to help solving the performance issue. Only the list and text fields are using
a local cache in this contribution: these fields are either intensive in terms
of SQL queries or CPU.

This is part of request #12245: Displaying an artifact with a lot of artifact links is slow

Change-Id: I26b4d2f4ecadb55dfbee6b245cefe35a6beb6452