request #12250: Tracker webhook payload can be incorrect if email notifications are disabled
The payload sent by a tracker webhook can indicate an artifact
has been created when in fact it's only an update.

The issue arises when the email notifications are not enabled because the email
notification task. The issue is caused by the fact the local changeset cache of
an artifact is incomplete after an update. The email notification task, when
executed, force fetch all the changesets to resolve inconsistencies in the cache.

This contribution ensures the cache is clean before executing other post action
tasks of an artifact.

Change-Id: I6474a01eddb56100391a9f5710df67bc33977da4